end_at int->double, as option -endpos accepts float seconds too patch by eviv bulgroz <ebulgroz@yahoo.com>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@8356 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ mencoder.c
@@ -286,7 +286,7 @@ static int interrupted=0;
 
 enum end_at_type_t {END_AT_NONE, END_AT_TIME, END_AT_SIZE};
 static enum end_at_type_t end_at_type = END_AT_NONE;
-static int end_at;
+static double end_at;
 
 static void exit_sighandler(int x){
     at_eof=1;
